Restructure LGFS demo claim generators
